Solve metabolite switch bug & Rename widget
